backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.18
pillow==11.0.0
numpy==2.1.3
python-dotenv==1.0.1
supabase==2.10.0
requests==2.32.3
pydantic==2.10.3
pydantic-settings==2.6.1
inference-sdk==0.9.0
google-generativeai==0.8.3
